+++ src/app/pages/hike-list/hike-list.component.ts
@@ -11,7 +11,7 @@ import { HikeSelectors } from 'subrepos/gtrack-common-ngx';
 import { LanguageService } from 'app/shared/services';
 
 import * as _ from 'lodash';
-import { SelectItem } from 'primeng/primeng';
+import { SelectItem, ConfirmationService } from 'primeng/primeng';
 
 @Component({
   selector: 'gt-hike-list',
@@ -28,7 +28,8 @@ export class HikeListComponent implements OnInit, OnDestroy {
   constructor(
     private _store: Store<State>,
     private _hikeSelectors: HikeSelectors,
-    private _title: Title
+    private _title: Title,
+    private _confirmationService: ConfirmationService
   ) {}
 
   ngOnInit() {
@@ -54,7 +55,12 @@ export class HikeListComponent implements OnInit, OnDestroy {
   }
 
   public deleteHike(hikeId: string) {
-    this._store.dispatch(new commonHikeActions.DeleteHikeProgram(hikeId));
+    this._confirmationService.confirm({
+      message: 'Are you sure that you want to delete?',
+      accept: () => {
+        this._store.dispatch(new commonHikeActions.DeleteHikeProgram(hikeId));
+      }
+    });
   }
 
   public translateDescription(description, field) {
